% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.blocksDS.R
\name{check.blocksDS}
\alias{check.blocksDS}
\title{Name imputation blocks}
\usage{
check.blocksDS(blocks, data, calltype = "type")
}
\arguments{
\item{prefix}{A character vector of length 1 with the prefix to
be using for naming any unnamed blocks with two or more variables.}
}
\value{
A named list of character vectors with variables names.
}
\description{
This helper function names any unnamed elements in the \code{blocks}
specification. This is a convenience function.
}
\details{
This function will name any unnamed list elements specified in
the optional argument \code{blocks}. Unnamed blocks
consisting of just one variable will be named after this variable.
Unnamed blocks containing more than one variables will be named
by the \code{prefix} argument, padded by an integer sequence
stating at 1.
}
\examples{
blocks <- list(c("hyp", "chl"), AGE = "age", c("bmi", "hyp"), "edu")
name.blocks(blocks)
}
\seealso{
\code{\link{mice}}
}
