% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.m.R
\name{check.m}
\alias{check.m}
\title{Partial Computation for Model Fitted Values}
\usage{
check.m(m)
}
\arguments{
\item{m}{is a list of the regression coefficients.}
}
\value{
a vector of fitted values.
}
\description{
Calculates the fitted values in each data node.
}
\details{
Considering \code{y} as a response variable and x as study variable, the fitted values are the y-values that
would expect for the given x-values according to the best-fitting straight line.
}
\section{Dependencies}{

\code{\link{getVarbyFormula}}
}

