% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.predictorMatrixDS.R
\name{checkpredictorMatrixDS}
\alias{checkpredictorMatrixDS}
\title{Creates a \code{predictorMatrix} argument}
\usage{
checkpredictorMatrixDS(predictorMatrix, data, blocks = NULL)
}
\arguments{
\item{data}{A \code{data.frame} with the source data}

\item{blocks}{An optional specification for blocks of variables in
the rows. The default assigns each variable in its own block.}
}
\value{
A matrix
}
\description{
This helper function creates a valid \code{predictMatrix}. The
\code{predictorMatrix} is an argument to the \code{mice} function.
It specifies the target variable or block in the rows, and the
predictor variables on the columns. An entry of \code{0} means that
the column variable is NOT used to impute the row variable or block.
A nonzero value indicates that it is used.
}
\seealso{
\code{\link{make.blocks}}
}
