% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.blocksDS.R
\name{construct.blocksDS}
\alias{construct.blocksDS}
\title{Construct blocks from \code{formulas} and \code{predictorMatrix}}
\usage{
construct.blocksDS(formulas = NULL, predictorMatrix = NULL)
}
\value{
A \code{blocks} object.
}
\description{
This helper function attempts to find blocks of variables in the
specification of the \code{formulas} and/or \code{predictorMatrix}
objects. Blocks specified by \code{formulas} may consist of
multiple variables. Blocks specified by \code{predictorMatrix} are
assumed to consist of single variables. Any duplicates in names are
removed, and the formula specification is preferred.
\code{predictorMatrix} and \code{formulas}. When both arguments
specify models for the same block, the model for the
\code{predictMatrix} is removed, and priority is given to the
specification given in \code{formulas}.
}
\examples{
form <- name.formulas(list(bmi + hyp ~ chl + age, chl ~ bmi))
pred <- make.predictorMatrix(nhanes[, c("age", "chl")])
construct.blocks(formulas = form, pred = pred)
}
\seealso{
\code{\link{make.blocks}}, \code{\link{name.blocks}}
}
