% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.formulasDS.R
\name{extend.formulas}
\alias{extend.formulas}
\title{Extends formula's with predictor matrix settings}
\usage{
extend.formulas(formulas, data, blocks, predictorMatrix = NULL,
  auxiliary = TRUE, include.intercept = FALSE, ...)
}
\arguments{
\item{auxiliary}{A logical that indicates whether the variables
listed in \code{predictors} should be added to the formula as main
effects. The default is \code{TRUE}.}

\item{include.intercept}{A logical that indicated whether the intercept
should be included in the result.}
}
\value{
A list of formula's
}
\description{
Extends formula's with predictor matrix settings
}
\keyword{internal}
