% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarbyFormula.R
\name{getVarbyFormula}
\alias{getVarbyFormula}
\title{Access Data from Study and Response Variables by \code{formula}.}
\usage{
getVarbyFormula(formula, subset = NULL, weight = NULL, family = NULL)
}
\arguments{
\item{formula}{a string character to be transformed as an object of class \code{\link[stats]{formula}}.}

\item{subset}{optional vector which specifies the subset of observations to be used in the analysis.}

\item{weight}{a list or numerical value for weigthed analysis.}

\item{family}{a string character with the name of the error distribution and link function to be used in the analysis.
If \code{family} is set to 'binomial', it transforms the vector of observed responses \code{y.var}
into \code{\link[base]{factors}}.
If \code{family} is NULL, it does not transform the data.}
}
\value{
A list with components:
\item{bind.x}{a data matrix with the independent variables and their respective values.}
\item{bind.y}{a data matrix, with the response variable and its respective values.}
}
\description{
Data access and transformations for dependent and independent variables in regressions and model checking analysis.
}
\details{
Internal function.
Commonly used by methods relationed to regression and model checking analysis.
}
\author{
Paula R. Costa e Silva
}
