% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarByName.R
\name{getVarByName}
\alias{getVarByName}
\title{Access Data from Study Variables by name.}
\usage{
getVarByName(x)
}
\arguments{
\item{x}{a string character: the names of the studies variables.}
}
\value{
a data matrix which each column is a study variable and its respecive values.
}
\description{
Access the values for a given study variable or a list of variables.
}
\details{
Internal function.
}
\author{
Paula R. Costa e Silva
}
