% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyNas.R
\name{identifyNas}
\alias{identifyNas}
\title{Identify missing values}
\usage{
identifyNas(execution)
}
\value{
the variables names with missing values.
}
\description{
...
}
\details{
...
}
\author{
Paula Silva, Rui Camacho
}
