% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.naDS.R
\name{is.naDS}
\alias{is.naDS}
\title{is.na() for DataSHIELD server}
\usage{
is.naDS(x)
}
\arguments{
\item{x}{an \R object to be tested: the default method for
    \code{is.na} and \code{anyNA} handle atomic vectors, lists,
    pairlists, and \code{NULL}.}
}
\value{
A logical object conforming to \code{x}
}
\description{
This function is a wrapper for \code{base::is.na()}.
}
\details{
If the length of input vector is less than the set filter
a missing value is returned.
}
\examples{
library(mice)
pat <- is.naDS(nhanes)
}
\seealso{
\code{\link[base]{is.na}}
}
\author{
Adapted from \url{https://github.com/gflcampos/dsMice/blob/master/R/notNaDS.R}
}
