% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miceDS.R
\name{miceDS}
\alias{miceDS}
\title{mice algorithm for DataSHIELD server}
\usage{
miceDS(data, ...)
}
\arguments{
\item{data}{A data frame or a matrix containing the incomplete data.
Missing values are coded as \code{NA}.}

\item{\dots}{Arguments passed down to \code{mice::mice()}}
}
\value{
An object of class \code{mids}
}
\description{
This function is a wrapper for \code{mice::mice()}.
}
\examples{
library(mice)
imp <- miceDS(nhanes, maxit = 0)
}
\seealso{
\code{\link[mice]{mice}}, \code{\link[=mids-class]{mids}}
}
