% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimp.R
\name{nimp}
\alias{nimp}
\title{Number of imputations per block}
\usage{
nimp(vars = NULL)
}
\value{
A numeric vector of length \code{length(blocks)} containing
the number of cells that need to be imputed within a block.
}
\description{
Calculates the number of cells within a block for which imputation
is requested.
}
\examples{
where <- is.na(nhanes)

# standard FCS
nimp(where)

# user-defined blocks

}
\seealso{
\code{\link{mice}}
}
