% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notNA.R
\name{notNA}
\alias{notNA}
\title{Negate}
\usage{
notNA(x)
}
\arguments{
\item{x}{parameter}
}
\value{
negation
}
\description{
Negate
}
\details{
if the length of input vector is less than the set filter
a missing value is returned.
}
