% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notNaDS.R
\name{notNaDS}
\alias{notNaDS}
\title{Checks what values are not missing}
\usage{
notNaDS(x)
}
\arguments{
\item{x}{atomic vector, list and pairlist to be tested}
}
\value{
a suitable index vector for use with x.
}
\description{
This fucntion is a wrapper for: !is.na()
}
\details{
if the length of input vector is less than the set filter
a missing value is returned.
}
