% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmmDS.R
\name{pmmDS}
\alias{pmmDS}
\title{Computes statistical mean of a vector}
\usage{
pmmDS(y, ry, x, wy = NULL, donors = 5, matchtype = 1L, ridge = 1e-05,
  ...)
}
\arguments{
\item{xvect}{a vector}
}
\value{
a numeric, the statistical mean
}
\description{
Calculates the mean value.
}
\details{
if the length of input vector is less than the set filter
a missing value is returned.
}
\author{
Gaye, A.
}
