% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.visitSequence.R
\name{make.visitSequence}
\alias{make.visitSequence}
\title{Creates a \code{visitSequence} argument}
\usage{
make.visitSequence(data = NULL, blocks = NULL)
}
\value{
Vector containing block names
}
\description{
This helper function creates a valid \code{visitSequence}. The
\code{visitSequence} is an argument to the \code{mice} function that
specifies the sequence in which blocks are imputed.
}
\examples{
make.visitSequence(nhanes)
}
\seealso{
\code{\link{mice}}
}
